requirements.txt
google-auth-oauthlib>=0.4.6
google-auth-httplib2==0.2.0
google-api-python-client>=2.0.0
supabase>=0.0.1
python-dotenv>=0.19.0
schedule==1.2.1
requests>=2.28.0
google-generativeai>=0.3.0
feedparser>=6.0.10
beautifulsoup4>=4.11.0 